backend/requirements.txt
asgiref==3.11.1
Django==4.2.30
django-environ==0.13.0
sqlparse==0.5.5
typing_extensions==4.15.0
psycopg2-binary
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-redis
redis
Pillow
xhtml2pdf
